move two files to src/
